configuration for Google Play where you need to have the native 16kb image

+++ platforms/android/android-sqlite-native-driver/native/SQLiteNative_JNI.c
@@ -189,7 +189,7 @@ Java_io_liteglue_SQLiteNative_sqlc_1db_1prepare_1st__JLjava_lang_String_2(JNIEnv
 
   free(_res);
 
-  return instance;
+  return (jlong)instance;
 }
